fix(pages): temporary vite.config.ts fix
`vite.config.ts` requires rollupOptions.input for every single
page in the site. The point of having the site have prog. pages
is to avoid having to do stuff like this. For now, these pages are
automatically created with `.gitlab-ci.yml`. I don't think this
is a good idea, but as long as it works...